+++ CHANGELOG.md
@@ -9,15 +9,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-* Add `get_orders_count` that returns the number of existing Shopify orders
+*
 
 ### Changed
 
-* Improve `list_orders` by making return all the existing Shopify orders
+*
 
 ### Fixed
 
-*
+* Fix `list_orders_a` and `_fetch_many` methods
 
 ## [0.6.0] - 2021-12-02